+++ R/orbital_speed.R
@@ -13,7 +13,9 @@
 #'
 #' @examples
 #' # Compute orbital speed for varying Hs
-#' S <- t(sapply(1:10, \(h) jonswap(h)$spec))
+#' S <- t(sapply(1:10, function(h) {
+#'   jonswap(h)$spec
+#' }))
 #' orb_speeds <- compute_orbital_speeds(S, rscd_freq, depth = 100, z = 10)
 #' plot(1:10, orb_speeds[, 1],
 #'   type = "l",
@@ -34,8 +36,8 @@ compute_orbital_speeds <- function(spec, freq, z = 0, depth = Inf, output_speeds
   }
 
   stopifnot(n_freq == length(freq))
-  stopifnot(all(z >= 0))
-  stopifnot(all(z <= depth))
+  stopifnot(z >= 0)
+  stopifnot(z <= depth)
 
   # Compute k efficiently when depth is discretized
   k <- outer(